Moved graphing to a better place in the file structure
