+++ .gitignore
@@ -39,3 +39,4 @@ nosetests.xml
 
 # Virtual environment
 venv
+data-dev.sqlite-journal
