+++ app/main/views.py
@@ -1,7 +1,8 @@
 from flask import render_template, redirect, url_for, abort, flash, request, jsonify
 from flask_login import login_required, current_user
-from app.static.graphing.data_analysis import get_user_data_df
-from app.static.graphing.graphs import get_categories, get_title_and_df_key_from_tab_value, get_graph_dict
+
+from app.main.graphing.graphs import get_categories, get_title_and_df_key_from_tab_value, get_graph_dict
+from app.main.graphing.data_analysis import get_user_data_df
 from app.static.survey.survey import labels, cleveland_tech_survey
 from . import main
 from .forms import EditSurveyForm
